+++ src/xch/providers/leaflet/index.ts
@@ -1,1 +1,0 @@
-export { LeafletProvider } from "./leaflet_provider";