+++ src/xch/providers/leaflet_rpc/index.ts
@@ -1,0 +1,1 @@
+export { LeafletRPCProvider } from "./leaflet_rpc_provider";